refactor: app wave 13 import boundaries, diplomacy row split, test densify (#4269)
Close narrow-logic and src/ import gaps across app/lib, promote UI-facing
symbols to domain barrels and narrow logic contracts, decompose diplomacy
row actions, dedupe region-map corner values, and merge diplomacy panel
part1/part2 tests into a scenario-table suite with shared fixtures.
