+++ SPEC/program/repo-lint.md
@@ -55,6 +55,7 @@
 | `tool/check_app_asset_image_decode_dedup.dart` | AST scan over `app/lib/**` enforcing the shared Flame asset-image decode helper (Refs #3699): the `ui.decodeImageFromList(...)` PNG-decode idiom may appear **only** in the canonical helper file `app/lib/features/game/flame/caches/asset_image_cache.dart` (which exposes `decodeImageAsset(...)`); every icon cache and tileset must route decode through that helper instead of re-inlining `rootBundle.load` -> `Completer` -> `decodeImageFromList`. Additionally, every concrete icon cache (a public class declared in a `*_icon_cache.dart` file whose name ends with `IconCache`) must declare `extends AssetImageCache` so the load lifecycle / `_isLoading`/`_isLoaded` flags / structured logging / decode fan-out stay single-sourced in the base; the `AssetImageCache` base file is exempt and tilesets adopt only the decode helper (not the cache base). Generated suffixes, `_test.dart` / `/test/` paths, and fixture dirs are skipped; matches use AST method-invocation / extends-clause checks so comments are ignored; rule `repo.app_asset_image_decode_dedup` |
 | `tool/check_app_event_handler_scope_logic_boundary.dart` | Enforce that `app/lib/core/services/app_event_handler/app_event_handler_scope.dart` has no direct import from `app/lib/features/game/logic/**`; rule `repo.app_event_handler_scope_logic_boundary` |
 | `tool/check_app_no_part_directives.dart` | Forbid new Dart `part` / `part of` directives under `app/lib/**` outside a shrink-only grandfather allowlist while wave-9 de-part slices land (mirrors orders/turn/setup no-part gates; Refs #4117) — see `SPEC/program/app-no-part-directives.md`; rule `repo.app_no_part_directives` |
+| `tool/check_app_lib_no_src_imports.dart` | Forbid `package:colonizethis_<domain>/src/` import directives under `app/lib/**`; UI must consume promoted symbols through domain barrels or narrow logic contract entrypoints (`turn_time_api.dart`, `civilian_intel_api.dart`, etc.); rule `repo.app_lib_no_src_imports` (Refs #4269) |
 | `tool/check_subscription_tracker.dart` | Enforce that `app/lib/features/**/*.dart` does not store subscriptions in raw `StreamSubscription` lists; use `SubscriptionTracker` for shared-lifecycle subscription cleanup; rule `repo.subscription_tracker` |
 | `tool/check_app_lib_no_suggest_work_orders.sh` | Forbid `suggestWorkOrders(` call sites under `app/lib/**` (per-unit UI uses `getAvailableWorkTargetsForUnit` / `getValidWorkOrderTileKeysWithVisibility`; Refs #2133); rule `repo.app_lib_no_broad_suggest_work_orders` |
 | `tool/check_flutter_action_pins.dart` | Enforce that every `.github/workflows/*.{yml,yaml}` step using `subosito/flutter-action@...` defines `with.flutter-version` with semantic version value **>= `3.41.9`**; rule `repo.flutter_action_pins` |
