+++ app/lib/features/game/flame/caches/per_player_work_target_selection_cache.dart
@@ -1,3 +1,3 @@
-export 'package:colonizethis_orders/src/orders/per_player_work_target_selection_cache.dart' show PerPlayerWorkTargetSelectionCache, WorkTargetSelectionPopulationStrategy, WorkTargetSelectionSnapshot;
+export 'package:colonizethis_orders/colonizethis_orders.dart' show PerPlayerWorkTargetSelectionCache, WorkTargetSelectionPopulationStrategy, WorkTargetSelectionSnapshot;
 // Implementation: colonizethis_logic (Refs #2277 B1 — shared worker/UI cache).
