+++ app/lib/features/game/screens/diplomacy/diplomacy_detail_screen.dart
@@ -3,7 +3,7 @@
 import 'package:colonizethis_diplomacy/colonizethis_diplomacy.dart'
     show diplomaticHistoryForPair, getOverture, greatPowerPowerScore;
 import 'package:colonizethis_economy/colonizethis_economy.dart' show PurchasedTileIndex;
-import 'package:colonizethis_logic/src/turn_to_year.dart' show turnToYear;
+import 'package:colonizethis_logic/turn_time_api.dart' show turnToYear;
 
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:flutter/material.dart';
