+++ app/lib/features/game/widgets/diplomacy/diplomacy_panel_rows.dart
@@ -1,6 +1,6 @@
 // Row model and builder for diplomacy UI. SPEC/ui/diplomacy-panel.md.
 
-import 'package:colonizethis_orders/src/orders/diplomatic_panel_actions.dart';
+import 'package:colonizethis_orders/colonizethis_orders.dart';
 
 import 'package:colonizethis_models/colonizethis_models.dart';
