+++ app/lib/features/game/widgets/units/civilian/civilian_units_panel_list.dart
@@ -1,6 +1,6 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_orders/colonizethis_orders.dart';
-import 'package:colonizethis_orders/src/orders/civilian_projected_tile.dart'
+import 'package:colonizethis_orders/colonizethis_orders.dart'
     show projectedCivilianTileKey;
 import 'package:colonizethis_world/colonizethis_world.dart';
