+++ app/lib/providers/map_view_provider_extraction.dart
@@ -1,6 +1,6 @@
 import 'package:colonizethis_app/core/utils/prefixed_id.dart';
 import 'package:colonizethis_data/colonizethis_data.dart';
-import 'package:colonizethis_economy/src/economy/economy_resource_constants.dart'
+import 'package:colonizethis_economy/economy_resource_constants_api.dart'
     show kMineralResourceIds;
 import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
