+++ app/test/diplomacy_panel_action_button_text_theme_test.dart
@@ -3,8 +3,7 @@
 // `AppThemes.editorialMonocle.textTheme` rather than a hard-coded
 // `const TextStyle(fontSize: 12)` literal.
 //
-// `DiplomacyActionButton` lives in `diplomacy_panel_row.dart` (extracted from
-// `diplomacy_panel.dart`, de-parted in #4117).
+// `DiplomacyActionButton` lives in `diplomacy_panel_row_actions.dart` (Refs #4269).
 //
 // SPEC:
 //  * `SPEC/ui/pixel-art-ui-catalog.md` § *Editorial-monocle palette* —
@@ -28,7 +27,7 @@ void main() {
   suppressLogsForTests();
 
   group(
-    'diplomacy_panel_row.dart DiplomacyActionButton routes caption through TextTheme '
+    'diplomacy_panel_row_actions.dart DiplomacyActionButton routes caption through TextTheme '
     '(Refs #2914 S7 regression guard)',
     () {
       late final String diplomacyPanelRowSource;
@@ -37,13 +36,13 @@ void main() {
         // `flutter test` runs from the package root (`app/`); the source
         // path is therefore relative to that working directory.
         final File source = File(
-          'lib/features/game/widgets/diplomacy/diplomacy_panel_row.dart',
+          'lib/features/game/widgets/diplomacy/diplomacy_panel_row_actions.dart',
         );
         expect(
           source.existsSync(),
           isTrue,
           reason:
-              'Expected `app/lib/features/game/widgets/diplomacy/diplomacy_panel_row.dart` '
+              'Expected `app/lib/features/game/widgets/diplomacy/diplomacy_panel_row_actions.dart` '
               'to exist; running directory is `${Directory.current.path}`.',
         );
         diplomacyPanelRowSource = source.readAsStringSync();
@@ -64,7 +63,7 @@ void main() {
           slotFallback.hasMatch(diplomacyPanelRowSource),
           isTrue,
           reason:
-              'Expected `DiplomacyActionButton.build` in `diplomacy_panel_row.dart` to '
+              'Expected `DiplomacyActionButton.build` in `diplomacy_panel_row_actions.dart` to '
               'resolve its caption text style via '
               '`theme.textTheme.bodySmall ?? const TextStyle(fontSize: 12)` '
               '(Refs #2914 S7). If the slot was intentionally changed, '
@@ -89,7 +88,7 @@ void main() {
           isEmpty,
           reason:
               'Found ${matches.length} `style: const TextStyle(fontSize: N)` '
-              'literal(s) in `diplomacy_panel_row.dart`. Each must route through '
+              'literal(s) in `diplomacy_panel_row_actions.dart`. Each must route through '
               '`Theme.of(context).textTheme.<slot> ?? const TextStyle(...)` '
               'per Refs #2914 S7 so font, weight, and colour flow from '
               '`AppThemes.editorialMonocle`. Offending occurrences: '
