+++ packages/colonizethis_economy/lib/economy_resource_constants_api.dart
@@ -1,0 +1,4 @@
+/// Narrow economy resource-constant surface for app UI callers (Refs #4269).
+library;
+
+export 'src/economy/economy_resource_constants.dart' show kMineralResourceIds;
