+++ packages/colonizethis_logic/lib/civilian_intel_api.dart
@@ -1,0 +1,6 @@
+/// Narrow civilian intel / missing-work surface for app UI callers (Refs #4269).
+library;
+
+export 'src/civilians/civilians_missing_work_orders.dart'
+    show CivilianMissingWorkOrderEntry, findCiviliansMissingWorkOrders;
+export 'src/civilians/spy_relocate_intel.dart';
