+++ packages/colonizethis_logic/lib/turn_time_api.dart
@@ -1,0 +1,4 @@
+/// Narrow turn-time mapping surface for app UI callers (Refs #4269).
+library;
+
+export 'src/turn_to_year.dart' show turnToYear;
