+++ packages/colonizethis_orders/lib/colonizethis_orders.dart
@@ -6,3 +6,6 @@ export 'src/debug_console/debug_console_regiments.dart';
 export 'src/debug_console/debug_console_ships.dart';
 export 'src/debug_console/debug_console_workers.dart';
 export 'src/orders/orders.dart';
+export 'src/orders/civilian_projected_tile.dart';
+export 'src/orders/diplomatic_panel_actions.dart';
+export 'src/orders/per_player_work_target_selection_cache.dart';
