+++ tool/ct_repo_lint_manifest.yaml
@@ -545,6 +545,13 @@ rules:
     runner: dart
     script: tool/check_app_narrow_logic_import.dart
 
+  - rule_id: repo.app_lib_no_src_imports
+    group: architecture
+    title: app/lib must not deep-import package src/ trees (Refs #4269 Slice B)
+    spec: SPEC/program/repo-lint.md
+    runner: dart
+    script: tool/check_app_lib_no_src_imports.dart
+
   - rule_id: repo.app_no_part_directives
     group: structure
     title: app lib must not add Dart part / part of directives outside shrink-only grandfather (Refs #4117)
